Switch file_size() function to return a long long rather than an int.  Some climate files we have are 2.6GB which overflows an integer.
Former-commit-id: cccd3ef96711752f0a4ad3c7cd78c420f5857831

+++ met/src/basic/vx_util/file_size.cc
@@ -32,7 +32,7 @@ using namespace std;
 ////////////////////////////////////////////////////////////////////////
 
 
-int file_size(const char * path)
+long long file_size(const char * path)
 
 {
 
@@ -53,7 +53,7 @@ if ( status < 0 )  {
    //  return the file size in bytes
    //
 
-return( (int) sbuf.st_size );
+return( (long long) sbuf.st_size );
 
 }
